Rename withQRHardwareAwareness file
