+++ app/components/UI/QRHardware/QRHardwareAwareness.tsx
