+++ app/components/UI/SignatureRequest/index.js
@@ -15,7 +15,7 @@ import Device from '../../../util/device';
 import Analytics from '../../../core/Analytics';
 import { ANALYTICS_EVENT_OPTS } from '../../../util/analytics';
 import { ThemeContext, mockTheme } from '../../../util/theme';
-import withQRHardwareAwareness from '../QRHardware/withQRHardwareAwareness';
+import withQRHardwareAwareness from '../QRHardware/QRHardwareAwareness';
 import QRSigningDetails from '../QRHardware/QRSigningDetails';
 
 const createStyles = (colors) =>
